+++ pkg/inject/secrets-provider_test.go
@@ -7,7 +7,6 @@ import (
 	"testing"
 )
 
-
 func TestSecretsProviderSidecarInjection(t *testing.T) {
 	var testCases = []injectionTestCase{
 		{
@@ -32,6 +31,17 @@ func TestSecretsProviderSidecarInjection(t *testing.T) {
 				"CONJUR_AUTHENTICATOR_ID": "my-authenticator-id",
 				"CONJUR_AUTHN_URL":        "https://conjur-oss.conjur-oss.svc.cluster.local/authn-k8s/my-authenticator-id",
 				"CONJUR_SSL_CERTIFICATE":  "-----BEGIN CERTIFICATE-----tVw0ZnjsOV2ZeIBRalX/72RplPzkmWKAw==\n-----END CERTIFICATE-----\n",
+			},
+		},
+		{
+			description:                         "SecretsProvider golden config",
+			annotatedPodTemplateSpecPath:        "./testdata/secrets-provider-annotated-pod.json",
+			expectedInjectedPodTemplateSpecPath: "./testdata/secrets-provider-mutated-pod.json",
+			env: map[string]string{
+				"conjurAccount":           "myConjurAccount",
+				"conjurApplianceUrl":      "https://conjur-oss.conjur-oss.svc.cluster.local",
+				"authnK8sAuthenticatorID": "my-authenticator-id",
+				"conjurSslCertificate":    "-----BEGIN CERTIFICATE-----tVw0ZnjsOV2ZeIBRalX/72RplPzkmWKAw==\n-----END CERTIFICATE-----\n",
 			},
 		},
 	}
@@ -67,6 +77,9 @@ func TestSecretsProviderSidecarInjection(t *testing.T) {
 			// Assert that the modified Pod template spec should equal the expected Pod
 			// template spec.
 			assert.JSONEq(t, string(expectedMod), string(mod))
+			for envVar, _ := range tc.env {
+				os.Unsetenv(envVar)
+			}
 		})
 	}
 }